Commit part of hibernate

+++ src/main/java/com/vironit/kazimirov/service/impl/SubsectionServiceImpl.java
@@ -1,5 +1,8 @@
 package com.vironit.kazimirov.service.impl;
 
+import com.vironit.kazimirov.entity.Good;
+import com.vironit.kazimirov.exception.CantDeleteElement;
+import com.vironit.kazimirov.fakedao.DaoInterface.GoodDao;
 import com.vironit.kazimirov.fakedao.DaoInterface.SubsectionDao;
 import com.vironit.kazimirov.entity.Subsection;
 import com.vironit.kazimirov.exception.RepeatitionException;
@@ -8,20 +11,30 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
+
 @Service
 public class SubsectionServiceImpl implements SubsectionService {
 
     @Autowired
     private final SubsectionDao subsectionDao;
+    private final GoodDao goodDao;
 
     @Autowired
-    public SubsectionServiceImpl(SubsectionDao subsectionDao) {
+    public SubsectionServiceImpl(SubsectionDao subsectionDao, GoodDao goodDao) {
         this.subsectionDao = subsectionDao;
+        this.goodDao = goodDao;
     }
 
     @Override
     public void addSubsection(Subsection subsection) throws RepeatitionException {
-        subsectionDao.addSubsection(subsection);
+        Optional<Subsection> subsection1 = Optional.ofNullable(subsectionDao.findSubsectionByName(subsection.getTitle()));
+        if (subsection1.isPresent()==false) {
+            subsectionDao.addSubsection(subsection);
+        } else {
+            throw new RepeatitionException("such subsection is present");
+        }
+
     }
 
     @Override
@@ -35,9 +48,19 @@ public class SubsectionServiceImpl implements SubsectionService {
     }
 
     @Override
-    public void deleteSubsection(int idSubsection) {
-        subsectionDao.deleteSubsection(idSubsection);
+    public void deleteSubsection(int idSubsection) throws CantDeleteElement {
+        Subsection subsection=subsectionDao.findSubsectionById(idSubsection);
+        List<Good> goods=goodDao.findBySubsection(subsection);
+        if (goods.size()==0) {
+            subsectionDao.deleteSubsection(idSubsection);
+        } else {
+            throw new CantDeleteElement("this subsection is using with goods");
+        }
+    }
 
+    @Override
+    public Subsection findSubsectionById(int idSubsection) {
+        return subsectionDao.findSubsectionById(idSubsection);
     }
 }
